requirements.txt
Flask==3.0.0
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
